Add drawntabletesting to pygrametl (#31)
Authors:
- Soeren Kejser Jensen <devel@kejserjensen.dk>
- Christian Thomsen <chr@cs.aau.dk>

Drawn Table Testing (DTT) makes testing of ETL flows easier. DTT
simplifies defining tests where the state of the data sources and/or
the rows in the database must be taken into consideration. With DTT it
is easy to define the preconditions and postconditions of tests by
drawing a table with rows using text. A text-based representation of a
database table is a Drawn Table (DT). The definition of a DT can
contain constraints on columns, variables that specify the
relationship between cells, and multiple asserts make it easy to
compare the contents of a DT to a database table. dttr allows tests to
be created without any Python programming, and formattable.py can be
integrated with a text editor to automate the formatting of DTs.