Resolve #27 by fixing mistakes in the documentation (#35)
Proofread the documentation and code examples (#35)

Author:
- Soeren Kejser Jensen <devel@kejserjensen.dk>

Reviewer:
- Christian Thomsen <chr@cs.aau.dk>

Add sales to the example for CachedDimension. Change csvfile to f everywhere.
Ensure a hyphen is always used in user-defined. Update the required Python
version. Normalize the code examples in the documentation. Proofread and
normalize documentation. Add missing references and fix missing links to the API
documentation. Update the changelog with information about the revised
documentation.