Prevent contamination of input dictionary
I noticed that scdensure for a SCD table manipulates the input row object. For example, the key attribute is added to the dictionary. This could lead to UNIQUE constraint conflicts when scdensure is called for another table where the key attribute names are the same, because the key attribute is preserved even though a new one should be assigned. If the user really wants the key inside the dict, he can assign it himself from the return value of the scdensure function: `row[keyatt] = dimTable.scdensure(row)`